requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==3.7.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.2.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.13.2
uvicorn==0.24.0
